Make several `CharacterPF2e` methods #private
